Upload and preveiw added

+++ database/migrations/2019_09_12_110121_create_exams_table.php
@@ -19,7 +19,7 @@ class CreateExamsTable extends Migration
             $table->string('file_name');
             $table->string('name');
             $table->bigInteger('views')->nullable();
-            $table->decimal('rating')->nullable();
+            $table->integer('rating')->nullable();
             $table->string('grade')->nullable();
             $table->timestamps();
         });
